to move NumberPicker to component.
