+++ client/js/components/NumberPicker.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Select from '../components/Select'
+import Select from './Select'
 import { AppContext } from '../context/AppContext'
 import PropTypes from 'prop-types'
