+++ client/js/domain/UserProfile.js
@@ -4,7 +4,7 @@ import { AppContext } from '../context/AppContext'
 import EditPwdForm from './EditPwdForm'
 import { setSuccessMessage } from '../helper/messageGenerator'
 import config from '../config'
-import NumberPicker from './NumberPicker'
+import NumberPicker from '../components/NumberPicker'
 import API from '../api'
 import Select from '../components/Select'
 import { SAVE_SUCCESS } from '../config/wordMap'
